requirements.txt
pandas
pyyaml
rasterio
pydantic>=2.0
plotly
